requirements.txt
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gym-notices==0.0.8
gymnasium==1.0.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
torch==2.6.0
typing_extensions==4.12.2
